Added confirmation checks

+++ src/del.cpp
@@ -46,8 +46,12 @@ int main(int argc, char** argv)
 
 	if(!onlyChkC && !onlyChkM && looksLikeVariable(vi.variable) && vi.variable.back() != '/')
 	{
+		std::string xref = (vi.xref != "" ? "["+vi.xref+"]" : "");
 		std::string value = getAppOutput(vi.variable).output[0];
 		bool deleted = false;
+
+		fprintf(stdout,"%s%sDelete variable \"%c%s%c/%s\" (value: \"%s\") [y/N]?%s ",TEXT_YELLOW,TEXT_BOLD,vi.scope.sigil,xref.c_str(),vi.evalType,vi.key.c_str(),value.c_str(),TEXT_NORMAL);
+		if(getchar() != 'y') return 0;
 
 		if(vi.property != "")
 		{
@@ -72,12 +76,15 @@ int main(int argc, char** argv)
 		}
 
 		vi.scope.save();
-		std::string xref = (vi.xref != "" ? "["+vi.xref+"]" : "");
 		if(deleted)	output(Warning,"Variable \"%c%s%c/%s\" (value: \"%s\") has been deleted.",vi.scope.sigil,xref.c_str(),vi.evalType,vi.key.c_str(),value.c_str());
 		else		output(Error,"Variable \"%c%s%c/%s\" does not exist to be deleted.",vi.scope.sigil,xref.c_str(),vi.evalType,vi.key.c_str());
 	}
 	else if(!onlyChkC && !onlyChkM && vi.variable.back() == '/')
 	{
+		std::string xref = (vi.xref != "" ? "["+vi.xref+"]" : "");
+		fprintf(stdout,"%s%sDelete variable set \"%c%s%c/%s\" (value: \"%s\") [y/N]?%s ",TEXT_YELLOW,TEXT_BOLD,vi.scope.sigil,xref.c_str(),vi.evalType,vi.key.c_str(),getSetStr(vi).c_str(),TEXT_NORMAL);
+		if(getchar() != 'y') return 0;
+
 		bool atLeastOneKeyRemoved = false;
 		struct RemovedKey
 		{
@@ -108,7 +115,6 @@ int main(int argc, char** argv)
 					break;
 			}
 
-			std::string xref = (vi.xref != "" ? "["+vi.xref+"]" : "");
 			if(rk.isRemoved)
 			{
 				output(Warning,"Variable \"%c%s%c/%s\" (value: \"%s\") has been deleted.",vi.scope.sigil,xref.c_str(),rk.type,k.c_str(),v.c_str());
@@ -133,6 +139,9 @@ int main(int argc, char** argv)
 		{
 			if(stringcasecmp(entry,c_to_be_deleted+".char") || std::filesystem::is_directory(campaign_path+entry)) continue;
 
+			fprintf(stdout,"%s%sDelete character \"%s\" [y/N]?%s ",TEXT_YELLOW,TEXT_BOLD,left(entry,entry.length()-5).c_str(),TEXT_NORMAL);
+			if(getchar() != 'y') return 0;
+
 			std::filesystem::remove(campaign_path+entry);
 			if(std::filesystem::exists(campaign_path+entry+".bak")) std::filesystem::remove(campaign_path+entry+".bak");
 			output(Warning,"Character file \"%s\" has been deleted.",(campaign_path+entry).c_str());
@@ -148,6 +157,10 @@ int main(int argc, char** argv)
 		for(const auto& entry : getDirectoryListing(campaigns_dir))
 		{
 			if(stringcasecmp(entry,m_to_be_deleted) || !std::filesystem::is_directory(campaigns_dir+entry)) continue;
+
+			fprintf(stdout,"%s%sDelete campaign \"%s\" [y/N]?%s ",TEXT_YELLOW,TEXT_BOLD,entry.c_str(),TEXT_NORMAL);
+			if(getchar() != 'y') return 0;
+
 			std::filesystem::remove_all(campaigns_dir+entry);
 			output(Warning,"Campaign directory \"%s\" has been deleted.",(campaigns_dir+entry).c_str());
 			return 0;
